refactor(web): break up workspace router into services (#535 follow-up 1)
Move the file ops / diff / search / git / agent-switch logic out of the
1422-LOC workspace/router.ts and behind service-tier seams. The router
now contains validation + delegation only — no node:fs, no
node:child_process, no @vscode/ripgrep imports.

- services/files-service.ts: file CRUD + path-traversal / .git guards
- services/search-service.ts: file-name fuzzy + ripgrep content search
- services/diff-service.ts: listBranches, getDiff/Summary/FileDiff, revertFile
- services/workspace-service.ts: gitPullByWorkspaceId, gitPushByWorkspaceId,
  gitCommit, generateCommitMessage, switchAgent

Router dropped from 1422 → 496 LOC (in the ballpark of queue/router.ts at
404 and tasks/router.ts at 298).

Refs: https://github.com/band-app/band/issues/535
